+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # 2024-07-01
 
+- [#3](https://github.com/mmomtchev/static-portable-python/issues/3), on Linux, avoid dependencies on `sqlite3.so` and `libreadline.so.8`
+
+# 2024-07-01
+
 - [#1](https://github.com/mmomtchev/static-portable-python/pull/1), on Linux and macOS, build with static versions of the compression libraries obtained from `conan`